Added metadata support to the post_publish.rb script
Users of the post_publish script will now be able to pass metadata to Opencast in the same way as the post_archive script. This implicitly fixes the repository documentation, which never specified that passing metadata only worked for the post_archive integration.

To that end, the metadata-related code was outsourced into three modules in the new folder `oc_modules`. Both post processing scripts now require the presence of this folder to function.

The post_publish script was expanded to use the new modules. Furthermore, the single web-request to `/ingest/addMediapackage{wID}` was replaced by multiple web-requests (using the 'rest-client' gem), as the aforementioned endpoint does not offer a form field for ACLs.