Log All Ingest Errors
Due to a misconfiguration, the script was failing to open a TCP
connection to Opencast which caused an exception and caused the script
to fail.

Unfortunately, this error was silenced and not logged at all, making it
extremely hard to find out what was going on.

This patch makes sure that all errors are logged.