Send real start and end times in post publish
The post_publish script has not been sending the
start and end times
of the recording to Opencast. It relied on that information being
in the metadata tag in the events.xml, where it simply
was not. This would result in Opencast displaying the ingest
time as the start time instead of the start time of
the recording.

This patch fixes that by employing the same mechanism the
post_archive script already uses, which is looking at the timestamps
of the first and last events in the events.xml, and then
using those values.