Livekit Audio Fix
Livekit produces a separate audio stream for webcams in BBB 3.
This audio stream is handled with a new event in the `events.xml`.
This commit makes sure that the livekit audio stream is also added to
the tracks and sent to opencast.

FYI: The opencast workflows are already capable of combining the additional audio
stream.